requirements.txt
beautifulsoup4==4.4.1
certifi==2020.4.5.1
chardet==3.0.4
html5lib==0.999
idna==2.9
pi==0.1.2
python-dateutil==2.8.1
requests==2.23.0
six==1.14.0
urllib3==1.25.9
python-dotenv==0.17.1